requirements.txt
flask-cors
numpy
scipy
matplotlib
torch
torchvision
Pillow
tqdm
lmdb
pandas
click
colorama
Flask
gunicorn
itsdangerous
Jinja2
MarkupSafe
Werkzeug
